backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==21.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.6
Pillow 
PyYAML==6.0
python-dotenv==1.0.0
django-import-export==3.2.0